requirements.txt
Django
djangorestframework
python-dotenv
pillow
requests
django-cors-headers
gunicorn